backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pydantic-settings
python-multipart
pymupdf
ollama
groq
jinja2
python-dotenv
playwright
python-jose[cryptography]
bcrypt
matplotlib
email-validator
psycopg2-binary
alembic
pg8000
